+++ src/animation_player.hpp
@@ -25,7 +25,7 @@ class AnimationPlayer {
    */
   AnimationPlayer(std::unordered_map<uint16_t, Animation const>&&
                       animations_per_tile_index);
-  uint16_t animatedTileIndexFor(uint16_t tile_index, size_t tick) const;
+  uint16_t animatedTileIndexFor(uint16_t tile_index, int64_t tick) const;
 
  private:
   std::unordered_map<uint16_t, Animation const> animations_per_tile_index_;
