+++ src/character.hpp
@@ -8,13 +8,13 @@ class Character {
  public:
   Character(Navigator const& navigator, Position const& initialPosition,
             Rectangle const& rectangle);
-  size_t height() const;
+  int64_t height() const;
   void moveBy(Vector const& desired_displacement);
   Position const& position() const;
   Rectangle const& rectangle() const;
-  size_t width() const;
-  size_t x() const;
-  size_t y() const;
+  int64_t width() const;
+  int64_t x() const;
+  int64_t y() const;
 
  private:
   Navigator const& navigator_;
